Move Context to remove circular dependency
