+++ client/src/components/Editor.tsx
@@ -7,7 +7,7 @@ import { AllStyledComponent } from '@remirror/styles/emotion';
 import { ProsemirrorNode } from 'prosemirror-suggest';
 
 import { debounce } from '../util/debounce';
-import { EditorContext } from '../pages/Remix';
+import { EditorContext } from '../pages/Context';
 
 const DEBOUNCE_SAVE_DELAY_MS = 250;
