+++ client/src/components/Module.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ResumeContext } from '../pages/Remix';
+import { ResumeContext } from '../pages/Context';
 
 interface Props {
   module: models.Module
@@ -15,11 +15,11 @@ class Module extends React.Component<Props, States> {
     }
     this.handleToggle = this.handleToggle.bind(this)
   }
-  // static contextType = ResumeContext
-  // declare context: React.ContextType<typeof ResumeContext>
-
+  static contextType = ResumeContext
+  context!: React.ContextType<typeof ResumeContext>
   // Toggle the module to be in the resume or not
   handleToggle() {
+    console.log(this.context.data)
   }
 
   render() {
