+++ client/src/pages/Context.tsx
@@ -1,0 +1,7 @@
+import React, { createContext } from "react"
+
+export const ResumeContext = createContext({
+    data: undefined,
+    mutate: () => {}
+  })
+  export const EditorContext = createContext((json: any) => { })
